requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
ascii_colors==0.11.4
attrs==25.3.0
audioop-lts==0.2.2
beautifulsoup4==4.14.2
bert-score==0.3.13
bitarray==3.7.1
boto3==1.40.53
botocore==1.40.53
Brotli==1.1.0
cached-property==2.0.1
cachetools==5.5.2
cbor2==5.7.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
ckzg==2.1.4
click==8.3.0
coloredlogs==15.0.1
colorlog==6.9.0
configparser==7.2.0
contourpy==1.3.3
cryptography==46.0.2
crytic-compile==0.3.10
cssselect2==0.8.0
cycler==0.12.1
cytoolz==1.0.1
dataclasses-json==0.6.7
datasets==4.4.1
dill==0.4.0
distro==1.9.0
doclayout_yolo==0.0.4
dotenv==0.9.9
eth-account==0.13.7
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
faiss-cpu==1.12.0
fast-langdetect==0.2.5
fastapi==0.119.0
fasttext-predict==0.9.2.4
ffmpy==0.6.3
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
google-ai-generativelanguage==0.9.0
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.38.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==5.49.1
gradio_client==1.13.3
gradio_pdf==0.0.22
groovy==0.1.2
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
hexbytes==1.3.1
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.11.0
jмесpath==1.0.1
joblib==1.5.2
json_repair==0.52.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.9
langchain-community==0.3.29
langchain-core==1.2.9
langchain-google-genai==2.1.12
langchain-huggingface==1.1.0
langchain-ollama==0.3.8
langchain-openai==1.1.8
langchain-text-splitters==0.3.11
langgraph==1.0.8
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.4.30
lazy_loader==0.4
lightrag-hku==1.4.9.3
loguru==0.7.3
lru-dict==1.3.0
magika==0.6.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mdurl==0.1.2
mineru==2.5.4
mineru_vl_utils==0.1.13
modelscope==1.31.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nano-vectordb==0.0.4.3
networkx==3.5
numpy==2.2.6
ollama==0.5.4
omegaconf==2.3.0
onnxruntime==1.23.1
openai==2.18.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.3
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.2
parsimonious==0.10.0
pdfminer.six==20250506
pdfplumber==0.11.7
pdftext==0.6.3
pillow==11.3.0
pipmaster==1.0.9
polars==1.34.0
polars-runtime-32==1.34.0
prettytable==3.16.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
py-cpuinfo==9.0.0
py-ecc==8.0.0
py-evm==0.12.1b1
py-solc-x==2.0.4
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.10.1
pydantic_core==2.41.5
pydub==0.25.1
pydyf==0.12.1
Pygments==2.19.2
PyMuPDF==1.26.4
pyparsing==3.2.5
pypdf==6.1.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.17.2
pypinyin==0.55.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
raganything==1.2.8
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rlp==4.1.0
robust-downloader==0.0.2
rsa==4.9.1
ruff==0.14.0
s3transfer==0.14.0
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.1.2
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
slither-analyzer==0.11.3
sniffio==1.3.1
solc-select==1.2.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
stringzilla==4.2.1
sympy==1.14.0
tenacity==9.1.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.10.4
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.1
tomlkit==0.13.3
toolz==1.0.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.56.2
trie==3.1.0
typer==0.19.2
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.214
ultralytics-thop==2.0.17
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.13
weasyprint==68.0
web3==7.13.0
webencodings==0.5.1
websockets==15.0.1
xlsxwriter==3.2.9
xxhash==3.5.0
yarl==1.20.1
z3-solver==4.15.3.0
zopfli==0.4.0
zstandard==0.25.0
